+++ src/base/grid.cpp
@@ -1,5 +1,5 @@
 #include <iostream>
-#include "grid.hpp"
+#include "base/grid.hpp"
 
 Grid::Grid(unsigned int line, unsigned int column, unsigned int mine_number) :
   line_{line},
